migrate `ButtonSubmitInvert`
