+++ packages/app/src/library/Form/Unbond/UnbondInput.tsx
@@ -9,7 +9,7 @@ import { useNetwork } from 'contexts/Network';
 import { useActiveAccounts } from 'contexts/ActiveAccounts';
 import { InputWrapper } from '../Wrappers';
 import type { UnbondInputProps } from '../types';
-import { ButtonSubmitInvert } from 'kits/Buttons/ButtonSubmitInvert';
+import { ButtonSubmitInvert } from 'ui-buttons';
 import { planckToUnitBn } from 'library/Utils';
 
 export const UnbondInput = ({
