+++ packages/ui-buttons/src/ButtonSubmit/index.module.scss
@@ -2,8 +2,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 .btnSubmit {
-  // @include btn-pulse;
-
   border: 1px solid var(--accent-color-primary);
   border-radius: var(--button-border-radius-small);
   background: var(--accent-color-primary);
