+++ packages/ui-buttons/src/index.tsx
@@ -9,5 +9,6 @@ export * from './ButtonPrimary';
 export * from './ButtonPrimaryInvert';
 export * from './ButtonSecondary';
 export * from './ButtonSubmit';
+export * from './ButtonSubmitInvert';
 
 export * from './ButtonText';
